backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
replicate==0.15.4
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
loguru==0.7.2